[quant] Fix the parts that were missing after initial migration
After the initial migration from `torch.quantization` to `torch.ao.quantization`, some of the files did not change.
This happened because the migration was done in parallel, and some of the files were landed while the others were still in the original location.
This is the last fix in the AO migration phase 1, which completely enables the ao.quantization namespace.

Test Plan:

`python test/test_quantization.py`

ghstack-source-id: 51a3f060789e4e672dec251c13540eb1a3fb6c6d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66058